Footern is now at the bottom on all pages.
