+++ src/app/(create-form-page)/form-details/[formid]/page.module.scss
@@ -111,6 +111,7 @@
     }
   }
 
-//   .form_responses_container {
-//   }
+  .form_responses_container {
+    margin-bottom: 5rem;
+  }
 }
