+++ src/app/(create-form-page)/form-details/deletion-success/page.module.scss
